requirements.txt
pytest==7.2.1
mypy==1.2.0
python-dotenv==0.21.0
requests==2.28.1
customtkinter==5.1.2